bog/pipeline/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argopy==0.1.9
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.12b0
blessings==1.7
cachetools==5.2.0
certifi==2021.10.8
cftime==1.5.2
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
coverage==6.2
cycler==0.11.0
dask==2022.1.0
debugpy==1.5.1
decorator==4.4.2
entrypoints==0.3
erddapy==1.2.0
executing==0.8.2
Fiona==1.8.20
flake8==4.0.1
frozenlist==1.3.0
fsspec==2022.1.0
geopandas==0.9.0
ghp-import==2.1.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
gsw==3.4.0
idna==3.3
importlib-metadata==4.11.3
ipdb==0.13.9
ipykernel==6.7.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.1
locket==0.2.1
mapclassify==2.4.3
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mccabe==0.6.1
mergedeep==1.3.4
mock==4.0.3
multidict==6.0.2
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
netCDF4==1.5.8
networkx==2.5.1
newrelic==7.2.4.171
numpy==1.23.3
packaging==21.3
pandas==1.4.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
protobuf==3.20.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pymdown-extensions==9.4
pyparsing==3.0.4
pyproj==3.3.0
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==22.3.0
requests==2.26.0
rsa==4.8
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
unicodecsv==0.14.1
urllib3==1.26.7
watchdog==2.1.7
wcwidth==0.2.5
xarray==0.20.2
yarl==1.7.2
zipp==3.6.0